add error state for plugin fetching
