+++ extensions/blocks/contact-form/components/jetpack-newsletter-integration-settings-consent-block.js
@@ -1,0 +1,71 @@
+/**
+ * External dependencies
+ */
+import { Button } from '@wordpress/components';
+import { __ } from '@wordpress/i18n';
+import { useCallback, useMemo } from '@wordpress/element';
+import { useDispatch, useSelect } from '@wordpress/data';
+import { createBlock } from '@wordpress/blocks';
+
+const useInsertConsentBlock = () => {
+	const selectedBlock = useSelect( select => select( 'core/block-editor' ).getSelectedBlock(), [] );
+	const { insertBlock } = useDispatch( 'core/block-editor' );
+
+	const insertConsentBlock = useCallback( async () => {
+		const buttonBlockIndex = ( selectedBlock.innerBlocks ?? [] ).findIndex(
+			( { name } ) => name === 'jetpack/button'
+		);
+		if ( buttonBlockIndex === -1 ) {
+			return;
+		}
+
+		const newConsentBlock = await createBlock( 'jetpack/field-consent' );
+		await insertBlock( newConsentBlock, buttonBlockIndex, selectedBlock.clientId, false );
+	}, [ insertBlock, selectedBlock.clientId, selectedBlock.innerBlocks ] );
+
+	return { insertConsentBlock };
+};
+
+const NoConsentBlockSettings = () => {
+	const { insertConsentBlock } = useInsertConsentBlock();
+
+	return (
+		<>
+			<p>
+				{ __(
+					'You’re already collecting email contacts. Why not make sure you have permission to email them too?',
+					'jetpack'
+				) }
+			</p>
+			<Button isSecondary onClick={ insertConsentBlock } style={ { marginBottom: '1em' } }>
+				{ __( 'Add email permission request', 'jetpack' ) }
+			</Button>
+			<br />
+		</>
+	);
+};
+
+const shouldHaveConsentBlockSelector = innerBlocks => {
+	const hasEmailBlock = innerBlocks.some( ( { name } ) => name === 'jetpack/field-email' );
+	const hasConsentBlock = innerBlocks.some( ( { name } ) => name === 'jetpack/field-consent' );
+	if ( hasEmailBlock ) {
+		return ! hasConsentBlock;
+	}
+	return false;
+};
+
+const ConsentBlockSettings = () => {
+	const selectedBlock = useSelect( select => select( 'core/block-editor' ).getSelectedBlock(), [] );
+	const shouldHaveConsentBlock = useMemo(
+		() => shouldHaveConsentBlockSelector( selectedBlock.innerBlocks ),
+		[ selectedBlock.innerBlocks ]
+	);
+
+	if ( shouldHaveConsentBlock ) {
+		return <NoConsentBlockSettings />;
+	}
+
+	return null;
+};
+
+export default ConsentBlockSettings;
